pyproject.toml
[project]
name = "kotiki-home-bot"
version = "0.0.1"
dependencies = [
    "click>=8",
    "adaptix == 3.0.0b5",
    "tinytuya>=1",
    "pyaml >= 24",
]

[tool.setuptools.packages.find]
include = ["sensors*"]
